"Separating Jump Variation from Integrated Volatility via Wavelet Analysis, with Applications to Assessing the Seasonal Fluctuations in Nitrate Concentration in River Water in Iowa" result reproducing
The following R codes are used to produce the results in the paper -
Separating Jump Variation from Integrated Volatility via Wavelet
Analysis, with Applications to Assessing the Seasonal Fluctuations in
Nitrate Concentration in River Water in Iowa by
Jin Meng, Keith Schilling, Kung-Sik Chan. To be more specific, the
results that can be reproduced by the following R codes are Figures 1-9
and Table 1-4 in the paper.