The R codes in the document below are used to produce the results in our submission.
The R codes in the section 1.3 are used to reproduce the numerical results in the sections 5 Sim- ulation Studies and 6 Nitrate Concentration Analysis of the submitted paper “Separating Jump Variation from Integrated Volatility via Wavelet Analysis, with Applications to Assessing the Seasonal Fluctuations in Nitrate Concentration in River Water in Iowa”. Please notice that the codes for 5 Simulation Studies would take a few hours to produce the result due to the large number of replication. The running time for the rest of the codes is rather short.